skills/demix/pyproject.toml
[project]
name = "clawsep"
version = "0.1.0"
description = "Audio source separation (vocals / instruments) via Demucs"
requires-python = ">=3.10"
dependencies = [
    "demucs>=4.0",
    "torchcodec",
    "ffmpeg-python",
]
